requirements.txt
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.29.2
diskcache==5.6.3
einops==0.8.0
exceptiongroup==1.2.1
fastapi==0.110.3
filelock==3.14.0
flash-attn==2.5.8
frozenlist==1.4.1
fsspec==2024.3.1
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.23.0
idna==3.7
interegular==0.3.3
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lark==1.1.9
llvmlite==0.42.0
lm-format-enforcer==0.9.8
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.0.8
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.550.52
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
outlines==0.0.34
packaging==24.0
prometheus_client==0.20.0
protobuf==5.26.1
psutil==5.9.8
py-cpuinfo==9.0.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
PyYAML==6.0.1
pyzmq==26.0.3
ray==2.20.0
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
rpds-py==0.18.0
safetensors==0.4.3
scipy==1.13.0
sentencepiece==0.2.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.12
tiktoken==0.6.0
tokenizers==0.19.1
torch==2.2.1
tqdm==4.66.2
transformers==4.40.1
triton==2.2.0
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
vllm==0.4.1
vllm_nccl_cu12==2.18.1.0.4.0
watchfiles==0.21.0
websockets==12.0
xformers==0.0.25
